Fix compiler warning
CXX    centerim5-BuddyList.o
../../src/BuddyList.cpp: In member function ‘virtual void BuddyList::Filter::Draw()’:
../../src/BuddyList.cpp:114:55: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ src/BuddyList.cpp
@@ -111,7 +111,7 @@ void BuddyList::Filter::Draw()
 
   int x = 0;
   x += area->mvaddstring(x, 0, realw - x, _("Filter: "));
-  if (parent->filter_buffer_onscreen_width <= realw - x) {
+  if (static_cast<int>(parent->filter_buffer_onscreen_width) <= realw - x) {
     // optimized simple case
     area->mvaddstring(x, 0, parent->filter_buffer);
   }
